Messed around with getting it to compile, probably going to go with wyvern v2 anyways

+++ contracts/wyvern/exchange/Exchange.sol
@@ -68,6 +68,8 @@ contract Exchange is ExchangeCore {
         return setOrderFill(hash, fill);
     }
 
+    // registries mapping will be checked in this call
+    // makes two executeOrder calls in ExchangeCore's atomicMatch
     function atomicMatch_(uint[16] memory uints, bytes4[2] memory staticSelectors,
         bytes memory firstExtradata, bytes memory firstCalldata, bytes memory secondExtradata, bytes memory secondCalldata,
         uint8[2] memory howToCalls, bytes32 metadata, bytes memory signatures)
@@ -75,10 +77,10 @@ contract Exchange is ExchangeCore {
         payable
     {
         return atomicMatch(
-            Order(address(uints[0]), address(uints[1]), address(uints[2]), staticSelectors[0], firstExtradata, uints[3], uints[4], uints[5], uints[6]),
-            Call(address(uints[7]), AuthenticatedProxy.HowToCall(howToCalls[0]), firstCalldata),
-            Order(address(uints[8]), address(uints[9]), address(uints[10]), staticSelectors[1], secondExtradata, uints[11], uints[12], uints[13], uints[14]),
-            Call(address(uints[15]), AuthenticatedProxy.HowToCall(howToCalls[1]), secondCalldata),
+            Order(address(uint160(uints[0])), address(uint160(uints[1])), address(uint160(uints[2])), staticSelectors[0], firstExtradata, uints[3], uints[4], uints[5], uints[6]),
+            Call(address(uint160(uints[7])), AuthenticatedProxy.HowToCall(howToCalls[0]), firstCalldata),
+            Order(address(uint160(uints[8])), address(uint160(uints[9])), address(uint160(uints[10])), staticSelectors[1], secondExtradata, uints[11], uints[12], uints[13], uints[14]),
+            Call(address(uint160(uints[15])), AuthenticatedProxy.HowToCall(howToCalls[1]), secondCalldata),
             signatures,
             metadata
         );
